+++ scripts/compatible.js
@@ -57,12 +57,12 @@ const compatible = {
 			'avif': 'image/avif',
 			'avifs': 'image/avif-sequence',
 		},
-		heic: {
+		/*heic: {
 			'heic': 'image/heic',
 			'heif': 'image/heif',
 			'heics': 'image/heic-sequence',
 			'heifs': 'image/heif-sequence',
-		},
+		},*/
 		special: {
 			'tbn': 'image/jpeg',
 		},
@@ -183,12 +183,12 @@ const compatibleMime = {
 			'image/avif': 'avif',
 			'image/avif-sequence': 'avifs',
 		},
-		heic: {
+		/*heic: {
 			'image/heic': 'heic',
 			'image/heif': 'heif',
 			'image/heic-sequence': 'heics',
 			'image/heif-sequence': 'heifs',
-		},
+		},*/
 	},
 	compressed: {
 		all: {},
@@ -315,7 +315,7 @@ compatible.image.all = [
 	...compatible.image.ico,
 	...compatible.image.webp,
 	...compatible.image.avif,
-	...compatible.image.heic,
+	// ...compatible.image.heic,
 ];
 compatible.image.blob = [ // This image formats requires conversion to Blob to be displayed
 
@@ -323,7 +323,7 @@ compatible.image.blob = [ // This image formats requires conversion to Blob to b
 compatible.image.sharp = [ // This image formats requires sharp (custom build) to be displayed
 	...compatible.image.jp2,
 	...compatible.image.jxl,
-	...compatible.image.heic,	
+	// ...compatible.image.heic,	
 ];
 compatible.image.convert = [ // This image formats requires conversion to PNG to be displayed
 	...compatible.image.jxr,
@@ -349,7 +349,7 @@ compatibleMime.image.all = [
 	...compatibleMime.image.ico,
 	...compatibleMime.image.webp,
 	...compatibleMime.image.avif,
-	...compatibleMime.image.heic,
+	// ...compatibleMime.image.heic,
 ];
 compatibleMime.image.blob = [ // This image formats requires conversion to Blob to be displayed
 
@@ -357,7 +357,7 @@ compatibleMime.image.blob = [ // This image formats requires conversion to Blob 
 compatibleMime.image.sharp = [ // This image formats requires sharp (custom build) to be displayed
 	...compatibleMime.image.jp2,
 	...compatibleMime.image.jxl,
-	...compatibleMime.image.heic,
+	// ...compatibleMime.image.heic,
 ];
 compatibleMime.image.convert = [ // This image formats requires conversion to PNG to be displayed
 	...compatibleMime.image.jxr,